# src/main/java/io/github/lyutovaanastasia/specification/domain/model/request/TypeRequest.java
package io.github.lyutovaanastasia.specification.domain.model.request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class TypeRequest {

    private String name;
    private String color;
}